// src/games/TicTacToe/models/PhaserComponentColor.ts
export default interface PhaserComponentColor {
  foreColor?: string | undefined;
  foreOverColor?: string | undefined;
  foreDownColor?: string | undefined;
  foreUpColor?: string | undefined;
  backColor?: number | undefined;
  backOverColor?: number | undefined;
  backDownColor?: number | undefined;
  backUpColor?: number | undefined;
  backAlpha?: number | undefined;
  borderColor?: number | undefined;
  borderOverColor?: number | undefined;
  borderDownColor?: number | undefined;
  borderUpColor?: number | undefined;
  borderAlpha?: number | undefined;
}